Added Death Scene and transition.
Click on health to decrease it by 10, click on button on Game Over
screen to restart game. In order for scene transition to work, both
Scene1 and DeathScene must be in the Build Settings. Additionally, I am
not committing any metadata, or anything in Temp, because those files
are seeming to be giving me some trouble when making the commit.