LEVEL GENERATOR!!! (  kind of :/  )
Simple baby version of the level generator.  Right now it only displays
a fake stone texture i got from the tutorial.  I should have it doing 1
line of grass but I haven't gotten that working yet.  Links to the
tutorials are at the top of the script if anyone (*cough* Kyle *cough*)
wants to take a look.  I'm too tired now to fix it but I will look at it
more tomorrow