Simple procedural desert is done!
Alpha version of the desert is done.  It generates mostly sand with
little dots of water and stone here and there.  If it gives you an all
desert landscape, that's not a bug; you just randomly got a shitty
section of the Perlin Noise plane.  DesertTerrain or whatever I called
it is sitting in the prefab folder, ready to use.