CameraControl and HUD changes
Started to work on a script to have the camera following the player.
Decided to stop midway as I realized the method could change depending
on which view we ultimately decide to go for.  Also changed a couple of
values in the HUD so resources are added by 10 instead of by 1 when
clicking on the corresponding button (for debug purposes).