Tiny change to terrain generator
I have it extracting the player's location in the world and the
generator's location in the world.  I still have to figure out how to
make it tell which block they are on.  I also added a "TileObject" class
that will store stuff like trees, stone, etc.  I don't know if we want
to use that class directly or have "class Tree : TileObject{ ... } ".
That remains to be determined next week.