Minor Bugfixes/Changes
Improved combat by speeding up enemies and reducing sliding on the
player, giving more of a "pop" to the knockback. Decreased the lifespan
of the weapon object to prevent the player from running into it. Tripled
the maximum number of enemies spawned and the range they can spawn in to
account for the larger map. Fixed the Weapon Pause bug.