More buildings
Added 2 more buildings. One is a resource generator that gives the
player a set amount of resources every time the player touches the
building (increases as time goes by). The other building is a WIP
attack tower that is supposed to shoot out projectiles at the enemy. I
was planning on making it very similar to how the WeaponScript works
for the player.