Shields added
Added in functionality for crafting defensive weapons. Also changed when
the game checks for best equipment, now occurs after crafting. If anyone
wants to add new items, add them in GameplayEvents Inventory and
Crafting, and account for them in PlayerStats. Currently spears are 1, 2
and shields are 11, 12, so offense items can be organized in 1-10 for
now, def items organized in 11 - 20. We can reorganize this when we add
more items.