Improved endgame + added wincondition
Enemies now start to converge on the player during endgame. Having
tested it, the player tends to become quickly overwhelmed, and it is
fairly difficult to survive, even with Healing House + Stone equipment,
and died the first time I tested it. Additionally added a second
winTimer that begins to count once the endgame begins. After a certain
amount of time, a win condition is triggered (which doesn't do anything
yet.)