Second Enemy Framework
Implemented the framework for a second enemy, changed the structure of
the enemy scripts to be able to apply to different enemies.  Enemy
currently doesn't move, behavior is still being implemented. Also,
changed the thresholds for terrain to give more varied landscape. Old
thresholds are still there commented out, so that change is easy to
revert.