Started implementing water collision
Strange bug, I think it spawns the tiles in the right location but the
collision box only seems to work when the player spawns on a lake and
pushes him off, and then seems to stop working.  I will update this
later.  I also moved the terrain generator in Scene1 to 0, 0, 0.5 for
simplicity, and started implementing a way to spawn the player somewhere
other than a lake