Lag Reduction + slight issue with water Collision
Altered the way Water Collision is handled. Instead of there being a
collision object, checks the player's position against the blocks[,]
array to see if player is on water. Lag absolutely annihilated. Only
issue is that collision is not precise for telling when the player is on
water. This is purely a math issue. If someone else (Cody) wants to look
at the code I added to the bottom of PlayerMovement.cs, go ahead. I'll
be looking at it again in a bit, but issue is largely fixed.