Game Controller
Created a game controller that swaps out the terrain from under you.
Right now bound to debug key "o" for obliterate.  Minor change to
PlayerMovementScript so it always knows where the new terrain is for
purposes of checking against water.  Game controller still needs to set
this up to timer events and display the countdown timer.