DodgeRoll/DimensionDoor
A minor commit, implemented the ability for the player to effectively
dimension door for a dodge ability once every few seconds, will be
getting in and making it so there's either a simple animation or a
camera delay so it doesn't make me nautious sometime on Tuesday. Press
Spacebar while moving in a direction to dodge in that direction.