HUD Inventory (now works!)
(Making the UI a prefab fixed most of my earlier issues.)Implemented a
preliminary inventory on the HUD. It currently shows
weather the player has found a better sword, shield, or boots. Planning
on possibly adding a popup message when the user mouses over for the
item's description. Also need to figure out how to display different
sprites when we eventually have different types of the same item.