SpawnPlayerSafely is now called
Found the issue that was preventing SpawnPlayerSafely() from being
called. I still spawn sometimes in the water, and I believe that's an
issue with the method or with blocks[,]. However, I'm not familiar
enough with that method to tinker with it yet. However, it can now be
properly tested/debugged by Cody now that it's actually being called
properly.