Fog of War Attempt
Hey guys, this is what I was caught up trying to make work for the last
few hours. It's an acceptable solution to implementing fog of war if I
could get the damn thing to work. However I don't enough about Unity's
shaders to refactor the thing from working on xz planes to working on xy
planes, probably something simple but I'm too fed up with it to continue
working on it right now, but I don't want to lose all the code so I'm
committing it without it being implemented into the build.