Implemented One-Use Item Slot
For some reason, it kept resetting the position on me (i wanted it to
be the last slot) but i’m too tired to figure out why it’s being so
stupid. I also started looking into the button but I need more info on
how to call a function upon clicking a GUI element. GameFlowController
has the stuff I need but I probably need to make another helper script
to speed up the process.