Endgame Object + Event added
Implemented basic endgame. When the difficulty reaches maximum, an
endgame tower is spawned generally near the player. When the player
presses "g" when near the object, it triggers endgame, with increased
enemy spawns (These numbers are very easy to tinker with for balancing).
Note, pressing G is just for our sake. We can make it something else for
demonstration. When the player survives 60 seconds, they are taken to a
win screen. I have not implemented the timer on the hud, nor the
navigation on the hud. I don't know if Krishna wants to do that a
specific way using his GUI stuff.