Endgame/game start screen fade + some fixes
Fixed a bunch of small bugs. Still encountering issues where Unity
hangs, likely related to the hazards.
Screen now fades from white, and to white, at the start and end of game,
respectively. This took longer than expected because the JS and C#
stopped playing nice for some reason for a while. I am now working on
audio cues for the ending, and audio/visual cues for the start of the
final difficulty and endgame/