Visual/Audio communication for endgame/win.
Added in some audio and visual cues to help communicate to the player
for the endgame. A button pops up when Difficulty = max pausing time
until they click it, explaining what to do now. After activating the
endgame, a sound plays every tick of the clock. And finally, there's a
sound that plays during endgame fade. Also added some placeholder
victory music to the endscene.