Added Enemy2 to Spawner
Enemy 2 now spawns alongside Enemy 1. We can later change this so that
Enemy 2 only spawns at higher difficulties. Note: Enemy 2 is difficult
to defeat with the slower speed. Also change the spawn rates/ranges so
that we won't run into the issue of the player wandering a long time
without encountering enemies.