Horrible Unity-hanging crash fixed.
Crash was located in SpawnHazard() in HazardControlScript, and its call
to isValidLocation(location). In the Flood apocalypse, isValidLocation
only returns true if it can fall in water. However, it can also only
fall in an area near the player. Therefore, if the player is nowhere
near water, and it attempts to drop a rock in the Flood apocalypse, it
will get permanently stuck in the do while loop, preventing the update
tick from finishing and causing Unity to hang.

The bug has been fixed by including a check that sees if the do while
loop has run over five times. If so, it exits the loop. This is
implemented until isValidLocation has a way of not freezing in that loop
in the eventuality that there is no valid place to spawn in range.