Difficulty increase transition
Added a transition for the increase in difficulty. The code is mainly in
CameraControl. In order to trigger it, press C. When we tie Difficulty
Increase to a timer, put the timer in CameraControl, and have it change
boolZoom to be 1. I will be adding sound effects later, I just want to
commit the visual component for now.