Weapon&Combat Tweaks, health bar fix
Rounded up the health bar to the nearest int so we dont have floating
points in the HUD. Tweaked player movement a bit. Added another attack
type. Now the standard left click attack is a straight hitbox instead of
a swing. The old swing can still be performed with a right click. I want
to figure out how to have the new hitbox appear anywhere around the
player instead of just 8 directions.