Tweaked-Out Combat
This combat's bitchin, yo! But on the real, I changed 2ndary attack to a
shield bash that does more than double the knockback but NO damage. It's
a great crowd-controlling mechanic, especially when there are a shit ton
of enemies in the higher difficulties. Made first attack a bit wider.
Adjusted boots speed.