Fixed EndGameObject spawning in Lava
The endgameobject should no longer spawn in Lava, unless the player is
deep in lava with no land nearby, in which case it's largely irrelevant
since the player's pretty doomed anyway in that case. Additionally, I
decreased the rate at which lava encroaches on the player. At the rate
it was at before, I was always ending up in lava with no land to go to.
Those thresholds might be in need of tweaking further.